use dedicated import store
